Style details for hierarchy Add hovering for infoviews. FIx bugs
Infoview hovering is still buggy but works in principle
