+++ infoview/server.ts
@@ -1,11 +1,34 @@
 import { Server, Transport, Connection, Event, TransportError } from 'lean-client-js-core';
-import { ToInfoviewMessage, FromInfoviewMessage } from '../src/typings';
+import { ToInfoviewMessage, FromInfoviewMessage, Config, Location } from '../src/typings';
 declare const acquireVsCodeApi;
 const vscode = acquireVsCodeApi();
 
 export function post(message: FromInfoviewMessage) { // send a message to the extension
+    console.log('posting to extension: ', message);
     vscode.postMessage(message);
 }
+
+export const PositionEvent: Event<Location> = new Event();
+export const ConfigEvent: Event<Config> = new Event();
+export const SyncPinEvent: Event<{pins: Location[]}> = new Event();
+
+window.addEventListener('message', event => { // messages from the extension
+    const message: ToInfoviewMessage = event.data; // The JSON data our extension sent
+    console.log('Received from extension:', message);
+    switch (message.command) {
+        case 'position':
+            PositionEvent.fire(message.loc);
+            break;
+        case 'on_config_change':
+            ConfigEvent.fire(message.config);
+            break;
+        case 'sync_pin':
+            SyncPinEvent.fire(message);
+            break;
+        // case 'copy_to_comment_request':
+        //     CopyToCommentEvent.fire(message);
+    }
+});
 
 class ProxyTransport implements Transport {
     connect(): Connection {
